Switched to tar for package extraction (changed download accordingly), many miscellaneous changes, extraction is now handled in /tmp and added a safeguard for wrong directories

+++ ogar-linux-script.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 echo "OGAR INSTALL, UPDATE AND UNINSTALL SCRIPT"
-echo "Make sure you have: nodejs, npm, bsdtar/unzip and wget/curl (for automatic downloads) installed!"
-echo "However, you may also download and extract master.zip manually."
+echo "Make sure you have: nodejs, npm, tar and wget/curl (for automatic downloads) installed!"
+echo "However, you may also download and extract master.tar.gz manually."
 echo "Place it in the same directory as the installer and name the folder Ogar-master."
 echo "----------------------------------------------------------------------------------------------------"
 if [ ! "$(id -u)" = 0 ]; then
@@ -19,71 +19,66 @@ fi
 
 echo "The Ogar server will be installed inside $2/ogar."
 echo "Do you wish to continue? (Y/N)"
-read yn
+read -r yn
 case $yn in
 		[Yy]* ) ;;
 		 * ) exit 1;;
 esac
 
+if [ ! -d "$2" ]; then
+		echo "$2 doesn't exist or is otherwise not accessible. Make sure you use absolute paths."
+		exit 1
+fi
+
 if grep "Arch Linux" /etc/*-release > /dev/null; then
 	echo "You are running Arch Linux. It is recommended to use the Ogar AUR package - https://aur4.archlinux.org/packages/ogar-git/"
 	echo "Do you wish to continue? (Y/N)"
-	read yn
+	read -r yn
 	case $yn in
 		[Yy]* ) ;;
 		* ) exit 1;;
 	esac
 fi
 if [ ! -d Ogar-master ]; then
-	if [ ! -f master.zip ]; then
-			echo "No local master.zip found, downloading with curl."
-			curl -O -L https://github.com/forairan/Ogar/archive/master.zip
+	if [ ! -f master.tar.gz ]; then
+			echo "No local master.tar.gz found, downloading with curl."
+			curl -O -L https://github.com/forairan/Ogar/archive/master.tar.gz
 	fi
-	if [ ! -f master.zip ]; then
-		echo "curl failed to download master.zip, trying wget."
-		wget https://github.com/forairan/Ogar/archive/master.zip
-			if [ ! -f master.zip ]; then
+	if [ ! -f master.tar.gz ]; then
+		echo "curl failed to download master.tar.gz, trying wget."
+		wget https://github.com/forairan/Ogar/archive/master.tar.gz
+			if [ ! -f master.tar.gz ]; then
 					echo "wget failed as well. Aborting!"
 					exit 1
 			fi
 	fi
-	echo "master.zip found!"
-	if [  -f /usr/bin/bsdtar ]; then
-			echo "BSDtar found. Using BSDtar to extract the archive."
-			bsdtar -xf master.zip
-		else
-			if [ -f /usr/bin/unzip ]; then
-				echo "unzip found. Using unzip to extract the archive."
-				unzip master.zip
-				else
-				echo "No .zip decompression tool found. Aborting!"
-				exit 1
-			fi
-	fi
+	echo "master.tar.gz found!"
+	echo "Extracting master.tar.gz to /tmp."
+	tar -xzf master.tar.gz -C /tmp
 fi
+echo "Removing master.tar.gz."
+rm master.tar.gz
+echo "Entering temporary directory."
+cd /tmp || exit 1
 echo "Organising and cleaning up the extracted files."
-rm -R Ogar-master/bin
-rm Ogar-master/Launch.bat
 rm Ogar-master/src/Start.bat
 rm Ogar-master/.gitignore
 echo "Copying the generated ogar folder to $2."
 cp -RTf Ogar-master "$2"/ogar
-echo "Removing master.zip"
-rm master.zip
 echo "Removing temporary files"
 rm -R Ogar-master
 
 echo "Creating ogar user and group if they don't exist"
-  if ! getent group "ogar" >/dev/null; then
+if ! getent group "ogar" >/dev/null; then
 	groupadd -r ogar
-  fi
-  if ! getent passwd "ogar" >/dev/null; then
+fi
+if ! getent passwd "ogar" >/dev/null; then
 	useradd -r -M -N -g ogar -d "$2"/ogar -s /usr/bin/nologin -c 'Ogar Server' ogar
-  fi
+fi
 
 echo "Installing ws module"
 rm -R /root/.npm
-cd "$2"/ogar
+cd "$2"/ogar || exit 1
 npm install ws
 
 echo "Symlinking gameserver.ini to /etc/ogar"
@@ -107,55 +102,50 @@ fi
 
 echo "The Ogar server inside $2/ogar will be updated."
 echo "Do you wish to continue? (Y/N)"
-read yn
+read -r yn
 case $yn in
 		[Yy]* ) ;;
 		 * ) exit 1;;
 esac
 
+if [ ! -f "$2/ogar/src/index.js" ]; then
+		echo "$2/ogar/src/index.js either way doesn't exist or isn't accesible. Are you sure this is an Ogar installation? Make sure you use absolute paths."
+		exit 1
+fi
+
 if [ ! -d Ogar-master ]; then
-	if [ ! -f master.zip ]; then
-			echo "No local master.zip found, downloading with curl."
-			curl -O -L https://github.com/forairan/Ogar/archive/master.zip
+	if [ ! -f master.tar.gz ]; then
+			echo "No local master.tar.gz found, downloading with curl."
+			curl -O -L https://github.com/forairan/Ogar/archive/master.tar.gz
 	fi
-	if [ ! -f master.zip ]; then
-		echo "curl failed to download master.zip, trying wget."
-		wget https://github.com/forairan/Ogar/archive/master.zip
-			if [ ! -f master.zip ]; then
+	if [ ! -f master.tar.gz ]; then
+		echo "curl failed to download master.tar.gz, trying wget."
+		wget https://github.com/forairan/Ogar/archive/master.tar.gz
+			if [ ! -f master.tar.gz ]; then
 					echo "wget failed as well. Aborting!"
 					exit 1
 			fi
 	fi
-	echo "master.zip found!"
-	if [  -f /usr/bin/bsdtar ]; then
-			echo "BSDtar found. Using BSDtar to extract the archive."
-			bsdtar -xf master.zip
-		else
-			if [ -f /usr/bin/unzip ]; then
-				echo "unzip found. Using unzip to extract the archive."
-				unzip master.zip
-				else
-				echo "No .zip decompression tool found. Aborting!"
-				exit 1
-			fi
-	fi
+	echo "master.tar.gz found!"
+	echo "Extracting master.tar.gz to /tmp."
+	tar -xzf master.tar.gz -C /tmp
 fi
+echo "Removing master.tar.gz."
+rm master.tar.gz
+echo "Entering temporary directory."
+cd /tmp || exit 1
 echo "Organising and cleaning up the extracted files."
-rm -R Ogar-master/bin
-rm Ogar-master/Launch.bat
 rm Ogar-master/src/Start.bat
 rm Ogar-master/.gitignore
 rm Ogar-master/gameserver.ini
 echo "Copying the generated ogar folder to $2."
 cp -RTf Ogar-master "$2"/ogar
-echo "Removing master.zip"
-rm master.zip
 echo "Removing temporary files"
 rm -R Ogar-master
 
 echo "Updating ws module"
 rm -R /root/.npm
-cd "$2"/ogar
+cd "$2"/ogar || exit 1
 npm install ws
 
 echo "Setting proper permissions"
@@ -174,27 +164,32 @@ if [ "$2" = "" ]; then
 		exit 1
 fi
 
+if [ ! -f "$2/ogar/src/index.js" ]; then
+		echo "$2/ogar/src/index.js either way doesn't exist or isn't accesible. Are you sure this is an Ogar installation? Make sure you use absolute paths."
+		exit 1
+fi
+
 echo "The ENTIRE $2/ogar folder will be DELETED."
 echo "Do you wish to continue? (Y/N)"
-read yn
+read -r yn
 case $yn in
 		[Yy]* ) ;;
 		* ) exit 1;;
 esac
 
 echo "Removing ogar user and group"
-  if getent passwd "ogar" >/dev/null; then
+if getent passwd "ogar" >/dev/null; then
 	userdel ogar > /dev/null
-  fi
-  if getent group "ogar" >/dev/null; then
+fi
+if getent group "ogar" >/dev/null; then
 	groupdel ogar >/dev/null
-  fi
+fi
 
 echo "Unlinking /etc/ogar"
 unlink /etc/ogar
 
 echo "Removing ws module"
-cd "$2"/ogar
+cd "$2"/ogar || exit 1
 npm uninstall ws
 
 echo "Removing the ENTIRE Ogar folder"
@@ -206,4 +201,3 @@ fi
 #If no install/update/uninstall parameter is specified
 echo "Please specify if you want to install, update or uninstall."
 exit 1
-
